+++ dart/geobase/lib/geobase.dart
@@ -67,6 +67,7 @@ export 'src/coordinates/scalable/scalable.dart';
 export 'src/coordinates/scalable/scalable2i.dart';
 
 // geodesy
+export 'src/geodesy/base/geodetic_arc_segment.dart';
 export 'src/geodesy/ellipsoidal/ellipsoidal.dart';
 export 'src/geodesy/ellipsoidal/ellipsoidal_vincenty.dart';
 export 'src/geodesy/spherical/spherical_great_circle.dart';