+++ dart/geobase/lib/geodesy.dart
@@ -15,6 +15,7 @@
 /// Usage: import `package:geobase/geodesy.dart`
 library geodesy;
 
+export 'src/geodesy/base/geodetic_arc_segment.dart';
 export 'src/geodesy/ellipsoidal/ellipsoidal.dart';
 export 'src/geodesy/ellipsoidal/ellipsoidal_vincenty.dart';
 export 'src/geodesy/spherical/spherical_great_circle.dart';